refactor: bind the embedder verdict to the value it measured
Resetting the probe from an effect set three pieces of state after
render, so the stale width showed for a frame and the bookkeeping needed
to spare the selection that had just started its own probe was the kind
that goes wrong quietly. `useEmbedderProbe` stores the outcome with the
value it was measured on and reads it back only while that still
matches; the one remaining effect aborts the request and sets nothing.

Its test hosted the field with a pinned `value` and a stubbed `onChange`,
modelling a parent that drops every edit. No caller does that, and it
hid whether the field reads its own value at all, so the host is now
controlled and the discard it made untestable is covered.

The budget metric cards move to `budget-cards.ts`: presentation shaping
that changes for different reasons than the arithmetic it reads, and
`budget.ts` was over its line cap.
